backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
opencv-python>=4.8.0
numpy>=1.26.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
google-generativeai>=0.3.0
google-cloud-vision>=3.4.0
aiofiles>=23.0.0
pytineye>=1.0.0
google-genai
pillow
python-dotenv
groq
sentence-transformers
flask
flask-cors
supabase>=2.0.0
